rm yaml load* funcs, using ZKGraphYaml
