+++ index.js
@@ -19,7 +19,6 @@ export { deposit } from "./api/deposit_bounty.js";
 export { verify } from "./api/verify.js";
 export { getRawReceipts, getBlockByNumber } from "./common/ethers_helper.js";
 export  * as Error from "./common/error.js"
-export {yamlhealthCheck} from './common/config_utils.js';
 export { waitTaskStatus } from './requests/zkwasm_taskdetails.js';
 export { TaskDispatch } from './api/task_dispatcher.js';
 export { zkwasm_setup } from './requests/zkwasm_setup.js';
