+++ tests/testsrc/zkgraph3.yaml
@@ -26,4 +26,4 @@ mapping:
 dataDestinations:
   - kind: ethereum
     network: sepolia
-    address: Your_Destination_Contract_Address_Here
+    address: "0x1B17C66e37CB33202Fd1C058fa1B97e36b7e517D"
